Closes #12 Auto Album
Just implemented All Content and here's why.

The front end is not going to look through every in memory album. It is
going ot act as any other album and just look to the backend for what's
in it. I only stubbed out the content of this auto album to be a few
pictures - not actually all the pictures seen elsewhere. You could
imagine in a full implementation, that album will actually respond with
the correct content supplied by a functioning backend.